entrypoint: add a "real" entrypoint
Because the way OCI/docker handle cmd/entrypoint is a little screwy, let's
add both cmd and entrypoint as passthrough options, and add a third
real_entrypoint, which clears the others and sets entrypoint. Should users
only care about setting the current entrypoint, and not inheriting anything
from other layers, this is the option to use.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>