take base layer inputs into account during caching
In particular, make sure that we process the base layer's imports and other
things, so that if something below this layer has changed, we also rebuild
this layer.

Since this recursively calls Lookup, it looks all the way until there is
not a "built" type layer.

Note that this switches the cache from being keyed on a structure hash to
being keyed on the entry name as it is in the stacker.yaml. This is of
course a paradigm shift, since when someone re-names the layer, they'll
have to regenerate it. However, they would have had to do that anyways,
since the imports are stored by layer name and not structure hash, and we
need to hash the imports in order to figure out the correctness of the
current layer. So, let's remove this illusion. Since people renaming layers
and still getting caching was cute but not vital or even common, this seems
like an acceptable loss.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>