fix intermediate snapshots
the problem here is really that umoci's metadata is tied to a manifest,
while we're trying to cache stacks of layers which do not depend on the
manifest at all.

so, let's do some hackery whackery to make sure that we always have a copy
of the mtree file and metadata that points to the manifest we're about to
use.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>