yaml: add a generate_labels field to yaml
The idea here is to be able to annotate the image config label with stuff
from the inside of the image (i.e. rpm -qa or similar). To do this, the
easiest way is to generate the list from inside the image, so let's add a
post-processing step to allow that.

However, let's make sure that this lives on a temporary snapshot, so any
additional packages or other fs stuff that was installed in order to
generate this metadata is not remembered.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>