storage: hoist Unpack() to a storage function
Previously, most of the layer-caching-unpack smarts lived in cmd/umoci.go.
However, that meant we didn't get these caching optimizations when using
squashfs, since that code wasn't exercised.

Additionally, most of that logic didn't really need to live inside a
subcommand. We need subcommands for doing the actual extracting, but
deciding which things to extract can be done in the parent process.

Finally, all of that logic is btrfs specific.

So lets refactor things so that: 1. we decide what to unpack before execing
the subcommand, 2. the subcommand is smart enough to unpack squashfs layers
as squashfs and tar as tar.

This again required us to split some of the userns logic out into its own
container package so that we could use it from both the btrfs package and
the main stacker package.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>