use rootless fseval everywhere
d841d429d95b ("storage: hoist Unpack() to a storage function") made it so
that our intermediate snapshot mtree generation code runs in the parent
process, instead of the child one. However, in unprivileged mode, this code
may not work on paths which are chmod 000, since the kernel only allows
root to open these files (even if the user owns them). Let's use the
rootless fseval instead of the passthrough default here, as it is smart
enough to chmod +700 the file, open it, and then chmod -700 after it is
opened.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>