unpriv-setup: do something reasonable for overlay mode
for overlay, we definitely don't want to mount a btrfs loopback, as that'll
slow things down making it all go through the loopback device.

However, we do still need the uid setup, so it's not entirely a no-op.

In addition, it seems reasonable to check the same things we check when
creating the overlay backend so that people can ensure their system really
can do the overlay stuff. This patch does that as well.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>