overlay: move to mounting rootfs using LXC
The problem here is that when running unprivileged, we need to be root in
a user namespace in order to mount overlayfs, which stacker itself will not
be. However, LXC is in this position and has overlay support, so we can use
it to mount the rootfs instead.

The prep work happened in previous patches, so this is mostly about ripping
our mount detection code out (and slightly re-working the way tar
extraction works).

As a bonus, this lets us get rid of a lot of code.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>